Use shorter config string in digest
Instead of including the full config in our digest we can use a memoized
hash of the string.  This reduces the total number of bytes that must be
processed in `ResultCache#file_checksum` and thus reduces its overall
runtime, since Digest::MD5's runtime is proportional to the number of
bytes processed.