Reduce bytes consumed by Digest::MD5.hexdigest
This uses the Digest library's file method to avoid reading the entire
file contents into memory.  Even though it doubles the actual number of
digest calls (one to digest the file contents, another to digest the
concatenated string), it results in about 30% better performance overall
because of a dramatically reduced bytesize of the outer string.