Halve stat syscalls in ResultCache#file_checksum
`ResultCache#file_checksum` needs to call `File.stat` (which is backed
by the `stat(2)` syscall) to ensure that file mode changes alter the
checksum.  But to generate a checksum the method also calls
ConfigStore#for, which in turn calls `File.directory?` on the same path,
incurring a redundant `stat(2)` syscall.  This commit adds a second
optional param to ConfigStore#for for use by callers that have already
called File.stat on the file and can thus save an extra syscall.  This
is not incredibly elegant but the call is in a hot path and this change
saves a noticable amount of time on large codebases.

A better option would be to split ConfigStore#for into a method that
accepts a directory path and one that accepts a file path so that it
wouldn't need to do the directory-or-file checking at all (callers
already know whether they are passing a dir or file).  I tested this and
found it to be even faster but it resulted in  a much larger diff and
some broken tests from brittle mocks that I couldn't figure out how to
fix, so I will go with the simpler option.