Use incremental digester
All stdlib digest classes support an incremental mode, where the data to
be digested can be fed incrementally using the `update` method.  One
advantage of this mode is that it can reduce the number and size of
allocated objects.  Using this scheme to pass the digester our
metadata and then the file contents improves overall performance by
about 20%.